make common exporter
