+++ packages/dd-trace/src/ci-visibility/exporters/test-worker/writer.js
@@ -23,11 +23,18 @@ class Writer {
   }
 
   _sendPayload (data) {
+    // ## Jest
     // Only available when `child_process` is used for the jest worker.
     // eslint-disable-next-line
     // https://github.com/facebook/jest/blob/bb39cb2c617a3334bf18daeca66bd87b7ccab28b/packages/jest-worker/README.md#experimental-worker
     // If worker_threads is used, this will not work
     // TODO: make it compatible with worker_threads
+
+    // ## Cucumber
+    // This reports to the test's main process the same way test data is reported by Cucumber
+    // See cucumber code:
+    // eslint-disable-next-line
+    // https://github.com/cucumber/cucumber-js/blob/5ce371870b677fe3d1a14915dc535688946f734c/src/runtime/parallel/run_worker.ts#L13
     if (process.send) { // it only works if process.send is available
       process.send([this._interprocessCode, data])
     }
