+++ packages/dd-trace/src/exporter.js
@@ -18,10 +18,8 @@ module.exports = name => {
     case exporters.AGENT_PROXY:
       return require('./ci-visibility/exporters/agent-proxy')
     case exporters.JEST_WORKER:
-      return require('./ci-visibility/exporters/jest-worker')
     case exporters.CUCUMBER_WORKER:
-      // for the moment we'll use the same, but we have to change this!
-      return require('./ci-visibility/exporters/jest-worker')
+      return require('./ci-visibility/exporters/test-worker')
     default:
       return inAWSLambda && !usingLambdaExtension ? require('./exporters/log') : require('./exporters/agent')
   }
